Update circle.yml: Add --release to `cargo test` to reduce CI run time.
By making both `cargo test` and `cargo run` to be release target,
they do not have to build rocksdb.so twice for debug and release
targets. Since building rocksdb.so will take ~8 minutes on a CircleCI
build machine and rocksdb.so was built 6 times in older builds, this
will give us huge savings.